phase382-D-warnings(quality): fix 2 substantive clang-tidy bugs from Batch 1
- test_image_readback.cpp:82 — bugprone-implicit-widening-of-multiplication-result
  Cast kW to uint64_t BEFORE multiply to avoid u32*u32 widening.

- VulkanDevice.cpp:3907 — cert-err34-c (atoi parse error handling)
  Replace std::atoi with std::strtol + end-pointer check for robust
  parsing of CD_VULKAN_DEVICE_INDEX environment variable.

Tests: 106/106 PASS. Clang-tidy warnings cleared at both sites.

+++ engine/render/rhi/tests/test_image_readback.cpp
@@ -79,7 +79,7 @@ TEST(ImageReadback, NullDeviceWritesZerosOnSuccess)
     constexpr std::uint32_t kW = 4;
     constexpr std::uint32_t kH = 4;
     // RGBA8Unorm: 4 bytes per texel.
-    constexpr std::uint64_t kBytes = kW * kH * 4;
+    constexpr std::uint64_t kBytes = static_cast<std::uint64_t>(kW) * kH * 4;
     constexpr std::uint64_t kOffset = 8;
 
     auto tex = make_texture(dev, kW, kH, cd::rhi::Format::kRGBA8Unorm);
